fix(ui): ensure ui-tab-routes render as expected
The UiTabRoutes component will redirect to the first tab in the
list if that's not an index route.

For this to work though, we can't short-circuit returning null
when activeTab is not set, as the component doesn't know which
tab is active until after that redirect.

We may be able to optimize how this works, but in order to make
the tabs render we'll have to get rid of this line.